✅ Q33H9D code: done
Task: 202604210900-Q33H9D
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: split task artifact schema into domain modules while preserving public exports and serialized compatibility | details: targeted schema tests, schemas check, typecheck, lint/format, and framework bootstrap passed.
