+++ packages/core/src/tasks/task-artifact-schema.shared.ts
@@ -1,0 +1,63 @@
+import { z } from "zod";
+import { zodToJsonSchema } from "zod-to-json-schema";
+
+export type JsonSchemaDocument = Record<string, unknown>;
+
+export const NON_EMPTY_STRING = z.string().min(1);
+export const ISO_UTC_TIMESTAMP = z.string().datetime({ offset: true });
+export const NULLABLE_NON_EMPTY_STRING = NON_EMPTY_STRING.nullable();
+export const NULLABLE_ISO_UTC_TIMESTAMP = ISO_UTC_TIMESTAMP.nullable();
+
+const zodToJsonSchemaSafe = zodToJsonSchema as (
+  schema: z.ZodTypeAny,
+  options: {
+    target: "jsonSchema7";
+    $refStrategy: "none";
+  },
+) => JsonSchemaDocument;
+
+export function buildJsonSchemaDocument(
+  schema: z.ZodTypeAny,
+  meta: {
+    $id: string;
+    title: string;
+    description?: string;
+  },
+): JsonSchemaDocument {
+  const generated = zodToJsonSchemaSafe(schema, {
+    target: "jsonSchema7",
+    $refStrategy: "none",
+  });
+  const { $schema: _schema, definitions: _definitions, ...rest } = generated;
+  return {
+    $schema: "http://json-schema.org/draft-07/schema#",
+    $id: meta.$id,
+    title: meta.title,
+    ...(meta.description ? { description: meta.description } : {}),
+    ...rest,
+  };
+}
+
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return !!value && typeof value === "object" && !Array.isArray(value);
+}
+
+export function formatSchemaErrors(label: string, issues: z.ZodIssue[] | undefined): string {
+  if (!issues || issues.length === 0) return `${label} schema validation failed`;
+  const rendered = issues.map((issue) => {
+    const path = issue.path.length > 0 ? `${label}/${issue.path.join("/")}` : label;
+    return `${path}: ${issue.message}`;
+  });
+  return `${label} schema validation failed: ${rendered.join("; ")}`;
+}
+
+export function schemaErrors<T>(label: string, schema: z.ZodType<T>, value: unknown): string[] {
+  const parsed = schema.safeParse(value);
+  return parsed.success ? [] : [formatSchemaErrors(label, parsed.error.issues)];
+}
+
+export function assertValid<T>(label: string, schema: z.ZodType<T>, value: unknown): T {
+  const parsed = schema.safeParse(value);
+  if (!parsed.success) throw new Error(formatSchemaErrors(label, parsed.error.issues));
+  return parsed.data;
+}
